requirements.txt
streamlit
pandas
plotly
scikit-learn
matplotlib
numpy
google-generativeai
python-dotenv